Moving BlindDeconv to EzPlug

+++ src/plugins/mitiv/myEzPlug/MyMetadata.java
@@ -4,8 +4,8 @@ public class MyMetadata {
     //Just a public object with all psf values inside
     public double dxy     = 64.5;
     public double dz      = 160;
-    public double nxy     = 256;
-    public double nz      = 128;
+    public int    nxy     = 256;
+    public int    nz      = 128;
     public double na      = 1.4;
     public double lambda  = 542;
     public double ni      = 1.518;
